requirements.txt
# --- Framework & Gemini LLM ---
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-cohere>=0.3.0
google-generativeai>=0.8.0

# --- Vector Database & Retrieval ---
faiss-cpu>=1.8.0
rank_bm25>=0.2.2

# --- Evaluation Framework ---
ragas>=0.2.0

# --- Data Ingestion & Processing ---
pandas>=2.2.0
beautifulsoup4>=4.12.0
selenium>=4.20.0
webdriver-manager>=4.0.0
tiktoken>=0.7.0

# --- Web Interface & Utilities ---
streamlit>=1.35.0
python-dotenv>=1.0.0
asyncio>=3.4.3